Added option to not normalize data on input and fixed public/private bug.
Two of the constructors had their public/private modifiers mixed up, just made some other stuff explicitly public too.

Also added the ability to not normalize data on input, with a default parameter set to normalize it. Not normalizing can be useful in unittests when dealing with a simple problem that you hand worked out, but I do not think it will be useful in practice.