Removed hypergradient from regularizations.
When tuning the hyper-parameters you should always test the generalization error, or the error measured over a validation data set compared to the error on the training set. This does not allow the direct evaluation of the hyper-gradient. In fact the gradient of the error (over any data set) is independent of the actual data and only depends on the regularization definition. As a result, minimizing its contribution to the overall error with gradients will always force it to a zero contribution. A numerical gradient to reduce the difference in the error over the training and validation data sets would be required to fine tune the hyper-parameters in order to maximize generalization of a network.